+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 * Fixed Gmail OAuth2 email delivery by exporting a synchronous Nodemailer transporter instead of a Promise
 * Fixed camera status WebSocket updates for the MongoDB 6 `findOneAndUpdate` return format
+* Fixed boolean environment variables treating the string `false` as enabled
 
 ### Chores